Update for newer RN and React Update Android app for latest React Native by creating MainApplication.java and moving the appropriate configuration over. Bump dependencies versions Bump release version number
No changes were made to core voice package internals.
