+++ VoiceTest/android/app/src/main/AndroidManifest.xml
@@ -11,6 +11,7 @@
         android:targetSdkVersion="22" />
 
     <application
+      android:name=".MainApplication"
       android:allowBackup="true"
       android:label="@string/app_name"
       android:icon="@mipmap/ic_launcher"
