+++ VoiceTest/package.json
@@ -1,13 +1,13 @@
 {
   "name": "VoiceTest",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
-    "react": "^0.14.8",
-    "react-native": "^0.25.1",
+    "react": "^15.3.2",
+    "react-native": "^0.34.1",
     "react-native-voice": "^0.1.0"
   }
 }
